+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes for Palette
 
+## 1.2.0 - 2022-02-07
+
+### Added
+- New `badgeCount` property to display Craft's badge count values
+- Add badge counts for primary navigation items
+
+### Updated
+- Move the Utility navigation badge count into the `badgeCount` property
+
 ## 1.1.0 - 2022-02-07
 
 ### Added